common, common/crypto: move keccak hasher pool into common/crypto
The pool, the Hasher type and their helpers lived in common/hasher.go, with
common/crypto forwarding to them. Nothing in package common used them, so move
them next to Keccak256 and drop the forwarders. Rename NewHasher to
NewKeccak256Hasher now that it sits beside the SHA-256 helpers.
