+++ node/shards/state_cache.go
@@ -27,6 +27,7 @@ import (
 	"github.com/holiman/uint256"
 
 	"github.com/erigontech/erigon/common"
+	"github.com/erigontech/erigon/common/crypto"
 	"github.com/erigontech/erigon/diagnostics/metrics"
 	"github.com/erigontech/erigon/execution/types/accounts"
 )
@@ -433,8 +434,8 @@ func (sc *StateCache) get(key btree.Item) (CacheItem, bool) {
 func (sc *StateCache) GetAccount(address []byte) (*accounts.Account, bool) {
 	AccRead.Inc()
 	var key AccountItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -462,8 +463,8 @@ func (sc *StateCache) HasAccountWithInPrefix(addrHashPrefix []byte) bool {
 // GetDeletedAccount attempts to retrieve the last version of account before it was deleted
 func (sc *StateCache) GetDeletedAccount(address []byte) *accounts.Account {
 	key := &AccountItem{}
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -484,8 +485,8 @@ func (sc *StateCache) GetDeletedAccount(address []byte) *accounts.Account {
 func (sc *StateCache) GetStorage(address []byte, incarnation uint64, location []byte) ([]byte, bool) {
 	StRead.Inc()
 	var key StorageItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -509,8 +510,8 @@ func (sc *StateCache) GetStorage(address []byte, incarnation uint64, location []
 // Second return value is true if such item is found
 func (sc *StateCache) GetCode(address []byte, incarnation uint64) ([]byte, bool) {
 	var key CodeItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -563,8 +564,8 @@ func (sc *StateCache) readQueuesLen() (res int) {
 // SetAccountRead adds given account to the cache, marking it as a read (not written)
 func (sc *StateCache) SetAccountRead(address []byte, account *accounts.Account) {
 	var ai AccountItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -604,8 +605,8 @@ func (sc *StateCache) GetStorageByHashedAddress(addrHash common.Hash, incarnatio
 // SetAccountRead adds given account address to the cache, marking it as a absent
 func (sc *StateCache) SetAccountAbsent(address []byte) {
 	var ai AccountItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -684,8 +685,8 @@ func (sc *StateCache) setWrite(item CacheItem, writeItem CacheWriteItem, del boo
 // SetAccountWrite adds given account to the cache, marking it as written (cannot be evicted)
 func (sc *StateCache) SetAccountWrite(address []byte, account *accounts.Account) {
 	var ai AccountItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -700,8 +701,8 @@ func (sc *StateCache) SetAccountWrite(address []byte, account *accounts.Account)
 // SetAccountDelete is very similar to SetAccountWrite with the difference that there no set value
 func (sc *StateCache) SetAccountDelete(address []byte) {
 	var ai AccountItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -714,8 +715,8 @@ func (sc *StateCache) SetAccountDelete(address []byte) {
 
 func (sc *StateCache) SetStorageRead(address []byte, incarnation uint64, location []byte, value []byte) {
 	var si StorageItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -732,8 +733,8 @@ func (sc *StateCache) SetStorageRead(address []byte, incarnation uint64, locatio
 
 func (sc *StateCache) SetStorageAbsent(address []byte, incarnation uint64, location []byte) {
 	var si StorageItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -749,8 +750,8 @@ func (sc *StateCache) SetStorageAbsent(address []byte, incarnation uint64, locat
 
 func (sc *StateCache) SetStorageWrite(address []byte, incarnation uint64, location []byte, value []byte) {
 	var si StorageItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -771,8 +772,8 @@ func (sc *StateCache) SetStorageWrite(address []byte, incarnation uint64, locati
 
 func (sc *StateCache) SetStorageDelete(address []byte, incarnation uint64, location []byte) {
 	var si StorageItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -792,8 +793,8 @@ func (sc *StateCache) SetStorageDelete(address []byte, incarnation uint64, locat
 
 func (sc *StateCache) SetCodeRead(address []byte, incarnation uint64, code []byte) {
 	var ci CodeItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -806,8 +807,8 @@ func (sc *StateCache) SetCodeRead(address []byte, incarnation uint64, code []byt
 
 func (sc *StateCache) SetCodeAbsent(address []byte, incarnation uint64) {
 	var ci CodeItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -819,8 +820,8 @@ func (sc *StateCache) SetCodeAbsent(address []byte, incarnation uint64) {
 func (sc *StateCache) SetCodeWrite(address []byte, incarnation uint64, code []byte) {
 	// Check if this is going to be modification of the existing entry
 	var ci CodeItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
@@ -837,8 +838,8 @@ func (sc *StateCache) SetCodeWrite(address []byte, incarnation uint64, code []by
 func (sc *StateCache) SetCodeDelete(address []byte, incarnation uint64) {
 	// Check if this is going to be modification of the existing entry
 	var ci CodeItem
-	h := common.NewHasher()
-	defer common.ReturnHasherToPool(h)
+	h := crypto.NewKeccak256Hasher()
+	defer crypto.ReturnHasherToPool(h)
 	//nolint:errcheck
 	h.Sha.Write(address)
 	//nolint:errcheck
